Update README.md.gotmpl
This edit copies the changes relevant to this file that were made last week as initially proposed. 

I did not delete the expression that has included the chart description for the introduction to the readme. If I am correct and that is what {{ template "chart.description" . }} does, perhaps it can be deleted now.

Thank you.